Start rearranging pulses into a subpackage
