+++ src/qibolab/pulses/shape.py
@@ -1,12 +1,7 @@
-"""Pulse and PulseSequence classes."""
-import copy
+"""PulseShape class."""
 import re
 from abc import ABC, abstractmethod
-from dataclasses import dataclass, fields
-from enum import Enum
-from typing import Optional
 
-import numpy as np
 from qibo.config import log
 from scipy.signal import lfilter
 
@@ -16,81 +11,6 @@ SAMPLING_RATE = 1
 Used for generating waveform envelopes if the instruments do not provide
 a different value.
 """
-
-
-class PulseType(Enum):
-    """An enumeration to distinguish different types of pulses.
-
-    READOUT pulses triger acquisitions. DRIVE pulses are used to control
-    qubit states. FLUX pulses are used to shift the frequency of flux
-    tunable qubits and with it implement two-qubit gates.
-    """
-
-    READOUT = "ro"
-    DRIVE = "qd"
-    FLUX = "qf"
-    COUPLERFLUX = "cf"
-
-
-class Waveform:
-    """A class to save pulse waveforms.
-
-    A waveform is a list of samples, or discrete data points, used by the digital to analogue converters (DACs)
-    to synthesise pulses.
-
-    Attributes:
-        data (np.ndarray): a numpy array containing the samples.
-    """
-
-    DECIMALS = 5
-
-    def __init__(self, data):
-        """Initialise the waveform with a of samples."""
-        self.data: np.ndarray = np.array(data)
-
-    def __len__(self):
-        """Return the length of the waveform, the number of samples."""
-        return len(self.data)
-
-    def __hash__(self):
-        """Hash the underlying data.
-
-        .. todo::
-
-            In order to make this reliable, we should set the data as immutable. This we
-            could by making both the class frozen and the contained array readonly
-            https://numpy.org/doc/stable/reference/generated/numpy.ndarray.flags.html#numpy.ndarray.flags
-        """
-        return hash(self.data.tobytes())
-
-    def __eq__(self, other):
-        """Compare two waveforms.
-
-        Two waveforms are considered equal if their samples, rounded to
-        `Waveform.DECIMALS` decimal places, are all equal.
-        """
-        return np.allclose(self.data, other.data)
-
-    def plot(self, savefig_filename=None):
-        """Plot the waveform.
-
-        Args:
-            savefig_filename (str): a file path. If provided the plot is save to a file.
-        """
-        import matplotlib.pyplot as plt
-
-        plt.figure(figsize=(14, 5), dpi=200)
-        plt.plot(self.data, c="C0", linestyle="dashed")
-        plt.xlabel("Sample Number")
-        plt.ylabel("Amplitude")
-        plt.grid(
-            visible=True, which="both", axis="both", color="#888888", linestyle="-"
-        )
-        if savefig_filename:
-            plt.savefig(savefig_filename)
-        else:
-            plt.show()
-        plt.close()
 
 
 class ShapeInitError(RuntimeError):
@@ -709,546 +629,3 @@ class Custom(PulseShape):
 
     def __repr__(self):
         return f"{self.name}({self.envelope_i[:3]}, ..., {self.envelope_q[:3]}, ...)"
-
-
-@dataclass
-class Pulse:
-    """A class to represent a pulse to be sent to the QPU."""
-
-    start: int
-    """Start time of pulse in ns."""
-    duration: int
-    """Pulse duration in ns."""
-    amplitude: float
-    """Pulse digital amplitude (unitless).
-
-    Pulse amplitudes are normalised between -1 and 1.
-    """
-    frequency: int
-    """Pulse Intermediate Frequency in Hz.
-
-    The value has to be in the range [10e6 to 300e6].
-    """
-    relative_phase: float
-    """Relative phase of the pulse, in radians."""
-    shape: PulseShape
-    """Pulse shape, as a PulseShape object.
-
-    See
-    :py: mod:`qibolab.pulses` for list of available shapes.
-    """
-    channel: Optional[str] = None
-    """Channel on which the pulse should be played.
-
-    When a sequence of pulses is sent to the platform for execution,
-    each pulse is sent to the instrument responsible for playing pulses
-    the pulse channel. The connection of instruments with channels is
-    defined in the platform runcard.
-    """
-    type: PulseType = PulseType.DRIVE
-    """Pulse type, as an element of PulseType enumeration."""
-    qubit: int = 0
-    """Qubit or coupler addressed by the pulse."""
-
-    def __post_init__(self):
-        if isinstance(self.type, str):
-            self.type = PulseType(self.type)
-        if isinstance(self.shape, str):
-            self.shape = PulseShape.eval(self.shape)
-        # TODO: drop the cyclic reference
-        self.shape.pulse = self
-
-    @classmethod
-    def flux(cls, start, duration, amplitude, shape, **kwargs):
-        return cls(
-            start, duration, amplitude, 0, 0, shape, type=PulseType.FLUX, **kwargs
-        )
-
-    @property
-    def finish(self) -> Optional[int]:
-        """Time when the pulse is scheduled to finish."""
-        if None in {self.start, self.duration}:
-            return None
-        return self.start + self.duration
-
-    @property
-    def global_phase(self):
-        """Global phase of the pulse, in radians.
-
-        This phase is calculated from the pulse start time and frequency
-        as `2 * pi * frequency * start`.
-        """
-        if self.type is PulseType.READOUT:
-            # readout pulses should have zero global phase so that we can
-            # calculate probabilities in the i-q plane
-            return 0
-
-        # pulse start, duration and finish are in ns
-        return 2 * np.pi * self.frequency * self.start / 1e9
-
-    @property
-    def phase(self) -> float:
-        """Total phase of the pulse, in radians.
-
-        The total phase is computed as the sum of the global and
-        relative phases.
-        """
-        return self.global_phase + self.relative_phase
-
-    @property
-    def id(self) -> int:
-        return id(self)
-
-    def envelope_waveform_i(self, sampling_rate=SAMPLING_RATE) -> Waveform:
-        """The envelope waveform of the i component of the pulse."""
-
-        return self.shape.envelope_waveform_i(sampling_rate)
-
-    def envelope_waveform_q(self, sampling_rate=SAMPLING_RATE) -> Waveform:
-        """The envelope waveform of the q component of the pulse."""
-
-        return self.shape.envelope_waveform_q(sampling_rate)
-
-    def envelope_waveforms(
-        self, sampling_rate=SAMPLING_RATE
-    ):  #  -> tuple[Waveform, Waveform]:
-        """A tuple with the i and q envelope waveforms of the pulse."""
-
-        return (
-            self.shape.envelope_waveform_i(sampling_rate),
-            self.shape.envelope_waveform_q(sampling_rate),
-        )
-
-    def modulated_waveform_i(self, sampling_rate=SAMPLING_RATE) -> Waveform:
-        """The waveform of the i component of the pulse, modulated with its
-        frequency."""
-
-        return self.shape.modulated_waveform_i(sampling_rate)
-
-    def modulated_waveform_q(self, sampling_rate=SAMPLING_RATE) -> Waveform:
-        """The waveform of the q component of the pulse, modulated with its
-        frequency."""
-
-        return self.shape.modulated_waveform_q(sampling_rate)
-
-    def modulated_waveforms(self, sampling_rate):  #  -> tuple[Waveform, Waveform]:
-        """A tuple with the i and q waveforms of the pulse, modulated with its
-        frequency."""
-
-        return self.shape.modulated_waveforms(sampling_rate)
-
-    def __hash__(self):
-        """Hash the content.
-
-        .. warning::
-
-            unhashable attributes are not taken into account, so there will be more
-            clashes than those usually expected with a regular hash
-
-        .. todo::
-
-            This method should be eventually dropped, and be provided automatically by
-            freezing the dataclass (i.e. setting ``frozen=true`` in the decorator).
-            However, at the moment is not possible nor desired, because it contains
-            unhashable attributes and because some instances are mutated inside Qibolab.
-        """
-        return hash(
-            tuple(
-                getattr(self, f.name)
-                for f in fields(self)
-                if f.name not in ("type", "shape")
-            )
-        )
-
-    def __add__(self, other):
-        if isinstance(other, Pulse):
-            return PulseSequence(self, other)
-        if isinstance(other, PulseSequence):
-            return PulseSequence(self, *other)
-        raise TypeError(f"Expected Pulse or PulseSequence; got {type(other).__name__}")
-
-    def __mul__(self, n):
-        if not isinstance(n, int):
-            raise TypeError(f"Expected int; got {type(n).__name__}")
-        if n < 0:
-            raise TypeError(f"argument n should be >=0, got {n}")
-        return PulseSequence(*([copy.deepcopy(self)] * n))
-
-    def __rmul__(self, n):
-        return self.__mul__(n)
-
-    def is_equal_ignoring_start(self, item) -> bool:
-        """Check if two pulses are equal ignoring start time."""
-        return (
-            self.duration == item.duration
-            and self.amplitude == item.amplitude
-            and self.frequency == item.frequency
-            and self.relative_phase == item.relative_phase
-            and self.shape == item.shape
-            and self.channel == item.channel
-            and self.type == item.type
-            and self.qubit == item.qubit
-        )
-
-    def plot(self, savefig_filename=None, sampling_rate=SAMPLING_RATE):
-        """Plots the pulse envelope and modulated waveforms.
-
-        Args:
-            savefig_filename (str): a file path. If provided the plot is save to a file.
-        """
-
-        import matplotlib.pyplot as plt
-        from matplotlib import gridspec
-
-        waveform_i = self.shape.envelope_waveform_i(sampling_rate)
-        waveform_q = self.shape.envelope_waveform_q(sampling_rate)
-
-        num_samples = len(waveform_i)
-        time = self.start + np.arange(num_samples) / sampling_rate
-        fig = plt.figure(figsize=(14, 5), dpi=200)
-        gs = gridspec.GridSpec(ncols=2, nrows=1, width_ratios=[2, 1])
-        ax1 = plt.subplot(gs[0])
-        ax1.plot(
-            time,
-            waveform_i.data,
-            label="envelope i",
-            c="C0",
-            linestyle="dashed",
-        )
-        ax1.plot(
-            time,
-            waveform_q.data,
-            label="envelope q",
-            c="C1",
-            linestyle="dashed",
-        )
-        ax1.plot(
-            time,
-            self.shape.modulated_waveform_i(sampling_rate).data,
-            label="modulated i",
-            c="C0",
-        )
-        ax1.plot(
-            time,
-            self.shape.modulated_waveform_q(sampling_rate).data,
-            label="modulated q",
-            c="C1",
-        )
-        ax1.plot(time, -waveform_i.data, c="silver", linestyle="dashed")
-        ax1.set_xlabel("Time [ns]")
-        ax1.set_ylabel("Amplitude")
-
-        ax1.grid(
-            visible=True, which="both", axis="both", color="#888888", linestyle="-"
-        )
-        ax1.axis([self.start, self.finish, -1, 1])
-        ax1.legend()
-
-        modulated_i = self.shape.modulated_waveform_i(sampling_rate).data
-        modulated_q = self.shape.modulated_waveform_q(sampling_rate).data
-        ax2 = plt.subplot(gs[1])
-        ax2.plot(
-            modulated_i,
-            modulated_q,
-            label="modulated",
-            c="C3",
-        )
-        ax2.plot(
-            waveform_i.data,
-            waveform_q.data,
-            label="envelope",
-            c="C2",
-        )
-        ax2.plot(
-            modulated_i[0],
-            modulated_q[0],
-            marker="o",
-            markersize=5,
-            label="start",
-            c="lightcoral",
-        )
-        ax2.plot(
-            modulated_i[-1],
-            modulated_q[-1],
-            marker="o",
-            markersize=5,
-            label="finish",
-            c="darkred",
-        )
-
-        ax2.plot(
-            np.cos(time * 2 * np.pi / self.duration),
-            np.sin(time * 2 * np.pi / self.duration),
-            c="silver",
-            linestyle="dashed",
-        )
-
-        ax2.grid(
-            visible=True, which="both", axis="both", color="#888888", linestyle="-"
-        )
-        ax2.legend()
-        # ax2.axis([ -1, 1, -1, 1])
-        ax2.axis("equal")
-        if savefig_filename:
-            plt.savefig(savefig_filename)
-        else:
-            plt.show()
-        plt.close()
-
-
-class PulseSequence(list):
-    """A collection of scheduled pulses.
-
-    A quantum circuit can be translated into a set of scheduled pulses
-    that implement the circuit gates. This class contains many
-    supporting fuctions to facilitate the creation and manipulation of
-    these collections of pulses. None of the methods of PulseSequence
-    modify any of the properties of its pulses.
-    """
-
-    def __add__(self, other):
-        """Return self+value."""
-        return type(self)(super().__add__(other))
-
-    def __mul__(self, other):
-        """Return self*value."""
-        return type(self)(super().__mul__(other))
-
-    def __repr__(self):
-        """Return repr(self)."""
-        return f"{type(self).__name__}({super().__repr__()})"
-
-    def copy(self):
-        """Return a shallow copy of the sequence."""
-        return type(self)(super().copy())
-
-    @property
-    def ro_pulses(self):
-        """A new sequence containing only its readout pulses."""
-        new_pc = PulseSequence()
-        for pulse in self:
-            if pulse.type == PulseType.READOUT:
-                new_pc.append(pulse)
-        return new_pc
-
-    @property
-    def qd_pulses(self):
-        """A new sequence containing only its qubit drive pulses."""
-        new_pc = PulseSequence()
-        for pulse in self:
-            if pulse.type == PulseType.DRIVE:
-                new_pc.append(pulse)
-        return new_pc
-
-    @property
-    def qf_pulses(self):
-        """A new sequence containing only its qubit flux pulses."""
-        new_pc = PulseSequence()
-        for pulse in self:
-            if pulse.type == PulseType.FLUX:
-                new_pc.append(pulse)
-        return new_pc
-
-    @property
-    def cf_pulses(self):
-        """A new sequence containing only its coupler flux pulses."""
-        new_pc = PulseSequence()
-        for pulse in self:
-            if pulse.type is PulseType.COUPLERFLUX:
-                new_pc.append(pulse)
-        return new_pc
-
-    def get_channel_pulses(self, *channels):
-        """Return a new sequence containing the pulses on some channels."""
-        new_pc = PulseSequence()
-        for pulse in self:
-            if pulse.channel in channels:
-                new_pc.append(pulse)
-        return new_pc
-
-    def get_qubit_pulses(self, *qubits):
-        """Return a new sequence containing the pulses on some qubits."""
-        new_pc = PulseSequence()
-        for pulse in self:
-            if pulse.type is not PulseType.COUPLERFLUX:
-                if pulse.qubit in qubits:
-                    new_pc.append(pulse)
-        return new_pc
-
-    def coupler_pulses(self, *couplers):
-        """Return a new sequence containing the pulses on some couplers."""
-        new_pc = PulseSequence()
-        for pulse in self:
-            if pulse.type is not PulseType.COUPLERFLUX:
-                if pulse.qubit in couplers:
-                    new_pc.append(pulse)
-        return new_pc
-
-    @property
-    def finish(self) -> int:
-        """The time when the last pulse of the sequence finishes."""
-        t: int = 0
-        for pulse in self:
-            if pulse.finish > t:
-                t = pulse.finish
-        return t
-
-    @property
-    def start(self) -> int:
-        """The start time of the first pulse of the sequence."""
-        t = self.finish
-        for pulse in self:
-            if pulse.start < t:
-                t = pulse.start
-        return t
-
-    @property
-    def duration(self) -> int:
-        """Duration of the sequence calculated as its finish - start times."""
-        return self.finish - self.start
-
-    @property
-    def channels(self) -> list:
-        """List containing the channels used by the pulses in the sequence."""
-        channels = []
-        for pulse in self:
-            if not pulse.channel in channels:
-                channels.append(pulse.channel)
-        channels.sort()
-        return channels
-
-    @property
-    def qubits(self) -> list:
-        """The qubits associated with the pulses in the sequence."""
-        qubits = []
-        for pulse in self:
-            if not pulse.qubit in qubits:
-                qubits.append(pulse.qubit)
-        qubits.sort()
-        return qubits
-
-    def get_pulse_overlaps(self):  # -> dict((int,int): PulseSequence):
-        """Return a dictionary of slices of time (tuples with start and finish
-        times) where pulses overlap."""
-        times = []
-        for pulse in self:
-            if not pulse.start in times:
-                times.append(pulse.start)
-            if not pulse.finish in times:
-                times.append(pulse.finish)
-        times.sort()
-
-        overlaps = {}
-        for n in range(len(times) - 1):
-            overlaps[(times[n], times[n + 1])] = PulseSequence()
-            for pulse in self:
-                if (pulse.start <= times[n]) & (pulse.finish >= times[n + 1]):
-                    overlaps[(times[n], times[n + 1])] += [pulse]
-        return overlaps
-
-    def separate_overlapping_pulses(self):  # -> dict((int,int): PulseSequence):
-        """Separate a sequence of overlapping pulses into a list of non-
-        overlapping sequences."""
-        # This routine separates the pulses of a sequence into non-overlapping sets
-        # but it does not check if the frequencies of the pulses within a set have the same frequency
-
-        separated_pulses = []
-        for new_pulse in self:
-            stored = False
-            for ps in separated_pulses:
-                overlaps = False
-                for existing_pulse in ps:
-                    if (
-                        new_pulse.start < existing_pulse.finish
-                        and new_pulse.finish > existing_pulse.start
-                    ):
-                        overlaps = True
-                        break
-                if not overlaps:
-                    ps.append(new_pulse)
-                    stored = True
-                    break
-            if not stored:
-                separated_pulses.append(PulseSequence([new_pulse]))
-        return separated_pulses
-
-    # TODO: Implement separate_different_frequency_pulses()
-
-    @property
-    def pulses_overlap(self) -> bool:
-        """Whether any of the pulses in the sequence overlap."""
-        overlap = False
-        for pc in self.get_pulse_overlaps().values():
-            if len(pc) > 1:
-                overlap = True
-                break
-        return overlap
-
-    def plot(self, savefig_filename=None, sampling_rate=SAMPLING_RATE):
-        """Plot the sequence of pulses.
-
-        Args:
-            savefig_filename (str): a file path. If provided the plot is save to a file.
-        """
-        if len(self) > 0:
-            import matplotlib.pyplot as plt
-            from matplotlib import gridspec
-
-            fig = plt.figure(figsize=(14, 2 * len(self)), dpi=200)
-            gs = gridspec.GridSpec(ncols=1, nrows=len(self))
-            vertical_lines = []
-            for pulse in self:
-                vertical_lines.append(pulse.start)
-                vertical_lines.append(pulse.finish)
-
-            n = -1
-            for qubit in self.qubits:
-                qubit_pulses = self.get_qubit_pulses(qubit)
-                for channel in qubit_pulses.channels:
-                    n += 1
-                    channel_pulses = qubit_pulses.get_channel_pulses(channel)
-                    ax = plt.subplot(gs[n])
-                    ax.axis([0, self.finish, -1, 1])
-                    for pulse in channel_pulses:
-                        num_samples = len(
-                            pulse.shape.modulated_waveform_i(sampling_rate)
-                        )
-                        time = pulse.start + np.arange(num_samples) / sampling_rate
-                        ax.plot(
-                            time,
-                            pulse.shape.modulated_waveform_q(sampling_rate).data,
-                            c="lightgrey",
-                        )
-                        ax.plot(
-                            time,
-                            pulse.shape.modulated_waveform_i(sampling_rate).data,
-                            c=f"C{str(n)}",
-                        )
-                        ax.plot(
-                            time,
-                            pulse.shape.envelope_waveform_i(sampling_rate).data,
-                            c=f"C{str(n)}",
-                        )
-                        ax.plot(
-                            time,
-                            -pulse.shape.envelope_waveform_i(sampling_rate).data,
-                            c=f"C{str(n)}",
-                        )
-                        # TODO: if they overlap use different shades
-                        ax.axhline(0, c="dimgrey")
-                        ax.set_ylabel(f"qubit {qubit} \n channel {channel}")
-                        for vl in vertical_lines:
-                            ax.axvline(vl, c="slategrey", linestyle="--")
-                        ax.axis([0, self.finish, -1, 1])
-                        ax.grid(
-                            visible=True,
-                            which="both",
-                            axis="both",
-                            color="#CCCCCC",
-                            linestyle="-",
-                        )
-            if savefig_filename:
-                plt.savefig(savefig_filename)
-            else:
-                plt.show()
-            plt.close()
